POC2/requirements.txt
Flask
Flask-SQLAlchemy
PyPDF2
pdfkit
werkzeug
Pillow
reportlab
PyMuPDF
fpdf